requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
click-plugins==1.1.1.2
colorama==0.4.6
dnspython==2.8.0
filelock==3.25.2
idna==3.11
nvdlib==0.8.3
python-dateutil==2.9.0.post0
python-whois==0.9.6
requests==2.33.1
requests-file==3.0.1
shodan==1.31.0
six==1.17.0
soupsieve==2.8.3
tabulate==0.10.0
tldextract==5.3.1
typing_extensions==4.15.0
urllib3==2.6.3
xlsxwriter==3.2.9
